requirements.txt
absl-py==0.1.13
astor==0.6.2
astroid==1.6.3
bleach==1.5.0
certifi==2018.1.18
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
gast==0.2.0
grpcio==1.10.1
gunicorn==19.7.1
h5py==2.7.1
html5lib==0.9999999
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.5
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.14.2
Pillow==5.1.0
protobuf==3.5.2.post1
pylint==1.8.4
PyYAML==3.12
scipy==1.0.1
six==1.11.0
SQLAlchemy==1.2.6
tensorboard==1.7.0
tensorflow==1.0.0
termcolor==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
appdirs==1.4.1
packaging==16.8
pyparsing==2.1.10
